% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{anova_tukey}
\alias{anova_tukey}
\title{ANOVA and post-hoc Tukey}
\usage{
anova_tukey(data, control, type)
}
\arguments{
\item{data}{ExpressionSet, Data object for which the variance will be analyzed.}

\item{control}{Character, The condition to which the contrasts are generated (the control would be most appropriate).}

\item{type}{all" or "control" The type of contrasts that will be generated.}
}
\value{
An ExpressionSet object containing FDR estimates of differential expression.
}
\description{
\code{anova_tukey} performs an Analysis of Variance fit (\code{\link{aov}}) and subsequent Tukey Honest Significant Differences analysis (\code{\link{TukeyHSD}}).
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_exprset <- into_exprset_expdesign(example_unique, columns, exp_design)

example_vsn <- norm_vsn(example_exprset)
example_filter <- miss_val_filter(example_vsn, thr = 0)
example_impute <- imputation_MSn(example_filter, fun = "MinProb")

example_anova_tukey <- anova_tukey(example_impute, "Ctrl", "control")
}
