% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cutoffs}
\alias{cutoffs}
\title{Denote significant proteins}
\usage{
cutoffs(data, alpha = 0.05, lfc = 1)
}
\arguments{
\item{data}{ExpressionSet, Data object which will be filtered for significant proteins.}

\item{alpha}{Integer, sets the threshold for the false discovery rate (FDR).}

\item{lfc}{Integer, sets the threshold for the log fold change (lfc).}
}
\value{
An ExpressionSet object containing a variable "sign" denoting the significant proteins.
}
\description{
\code{cutoffs} denotes significant proteins based on defined cutoffs.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
example_exprset <- into_exprset(example_unique, columns)
example_vsn <- norm_vsn(example_exprset)
example_filter <- miss_val_filter(example_vsn, thr = 0)
example_impute <- imputation_MSn(example_filter, fun = "QRILC")

example_lm <- linear_model(example_impute, "Con_", "control")
example_sign <- cutoffs(example_lm, alpha = 0.05, lfc = 1)
significant_proteins <- example_sign[fData(example_sign)$sign == "+"]
nrow(significant_proteins)
}
