% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DiUbi_ExpDesign}
\alias{DiUbi_ExpDesign}
\title{Experimental design of the DiUbi dataset}
\format{A data.frame with 30 observations and 3 variables:
\describe{
  \item{label}{Label names}
  \item{condition}{Experimental conditions}
  \item{replicate}{Replicate number}
}}
\source{
Zhang, Smits, van Tilburg, et al (2017). An interaction landscape of ubiquitin signaling. Molecular Cell:
  \url{http://www.cell.com/molecular-cell/fulltext/S1097-2765(17)30004-7}
}
\usage{
DiUbi_ExpDesign
}
\value{
A data.frame with the DiUbi experimental design
}
\description{
This dataset annotates 30 different samples in 10 conditions and 3 replicates.
}
\keyword{datasets}
