% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{ends}
\alias{ends}
\title{dep::ends()}
\usage{
ends(project_root = getwd(), overwrite = FALSE, ...)
}
\arguments{
\item{project_root}{project root folder}

\item{overwrite}{overwrite an existing DESCRIPTION file?}

\item{...}{further arguments passed to \code{dsc$set}}
}
\value{
NULL; but side effect is the creation of a DESCRIPTION file within the project root folder
}
\description{
dep::ends()
}
\details{
Finds and documents (in a DESCRIPTION file) all R packages required to run all R code in a project.  The version of each package is taken from versions installed in the user's library (.libPaths()) and the current used version of R is also recorded.
}
