% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExpDesign_UbIA_MS}
\alias{ExpDesign_UbIA_MS}
\title{Experimental design of the UbIA_MS dataset}
\format{A data.frame with 12 observations and 3 variables:
\describe{
  \item{label}{Uniprot IDs}
  \item{sample}{Uniprot IDs of major protein(s) in the protein group}
  \item{replicate}{Full protein names}
}}
\source{
Zhang, Smits, van Tilburg, et al (2017). An interaction landscape of ubiquitin signaling. Molecular Cell
  \url{http://www.cell.com/molecular-cell/fulltext/S1097-2765(17)30004-7}
}
\usage{
ExpDesign_UbIA_MS
}
\description{
This dataset annotates the 12 different labels in 4 samples and 3 replicates.
}
\keyword{datasets}
