% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_missval}
\alias{filter_missval}
\title{Filter on missing values}
\usage{
filter_missval(data, thr = 0)
}
\arguments{
\item{data}{SummerizedExperiment, Data object which will be filtered.}

\item{thr}{Integer, sets the threshold for the allowed number of missing values per condition.}
}
\value{
An filtered SummerizedExperiment object.
}
\description{
\code{filter_missval} filters an SummerizedExperiment object based on missing values.
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)

example_stringent_filter <- filter_missval(example_se, thr = 0)
example_less_stringent_filter <- filter_missval(example_se, thr = 1)
}
