% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_df_long}
\alias{get_df_long}
\title{Generate a long data.frame from a SummarizedExperiment}
\usage{
get_df_long(se)
}
\arguments{
\item{se}{SummarizedExperiment,
Proteomics dataset.}
}
\value{
A data.frame object
containing all data in a wide format,
where each row represents a single measurement.
}
\description{
\code{get_df_long} generate a wide data.frame from a SummarizedExperiment.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.01)

diff <- test_diff(imputed, "Ctrl", "control")
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)

long <- get_df_long(dep)
colnames(long)
}
