% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_df_wide}
\alias{get_df_wide}
\title{Generate a wide data.frame from a SummarizedExperiment}
\usage{
get_df_wide(se)
}
\arguments{
\item{se}{SummarizedExperiment,
Proteomics data (output from \code{\link{make_se}()} or
\code{\link{make_se_parse}()}).}
}
\value{
A data.frame object
containing all data in a wide format,
where each row represents a protein.
}
\description{
\code{get_df_wide} generate a wide data.frame from a SummarizedExperiment.
}
\examples{
# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter, normalize and impute missing values
filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.01)

# Test for differentially expressed proteins
diff <- test_diff(imputed, "Ctrl", "control")
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)

# Get a wide data.frame
wide <- get_df_wide(dep)
colnames(wide)
}
