% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_prefix}
\alias{get_prefix}
\title{Obtain the longest common prefix}
\usage{
get_prefix(words)
}
\arguments{
\item{words}{Character vector,
A list of words.}
}
\value{
A character vector containing the prefix.
}
\description{
\code{get_prefix} returns the longest common prefix
of the supplied words.
}
\examples{
# Load example
data <- UbiLength
columns <- grep("LFQ.", colnames(data))

# Get prefix
names <- colnames(data[, columns])
get_prefix(names)
}
