% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{get_proj_deps}
\alias{get_proj_deps}
\title{Get packages required to run all R code in a project}
\usage{
get_proj_deps(root = ".")
}
\arguments{
\item{root}{project root folder to recursively search for R and Rmd files}
}
\value{
a vector of package names
}
\description{
Get packages required to run all R code in a project
}
\details{
Finds all R and Rmd files in a project and uses \code{\link{get_deps}}
    to find all R packages required for the code to run
}
