% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{get_stoichiometry}
\alias{get_stoichiometry}
\title{Relative stoichiometry}
\usage{
get_stoichiometry(data, ibaq, contrast, bait, level = 1)
}
\arguments{
\item{data}{SummarizedExperiment,
Proteomics dataset in which differential enriched proteins
are annotated by \code{\link{add_rejections}}.}

\item{ibaq}{Data.frame,
iBAQ table generated by \code{\link{merge_ibaq}}.}

\item{contrast}{Character,
The specific contrast for which to calculate the stoichiometries.}

\item{bait}{Character,
The name of the protein to which all other proteins will be scaled.}

\item{level}{Numerical,
The level to which the bait will be scaled.}
}
\value{
A data.frame with relative stoichiometry values.
}
\description{
\code{get_stoichiometry} calculates the relative stoichiometries of
all differentially enriched proteins.
}
\examples{
# load data
data <- GFPip
exp_design <- GFPip_ExpDesign
peptides <- GFPip_pep

# Make unique names and filter for reverse and contaminant hits
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs')
data_filt <- data_unique[data_unique$Reverse != '+' & data_unique$Contaminant != '+',]

# Make a SummarizedExperiment and perform filtering on missing values,
# variance stabilization and imputation on this object
cols <- grep('LFQ', colnames(data_filt))
se <- make_se(data_filt, cols, exp_design)
filt <- filter_missval(se)
norm <- normalize_vsn(filt)
imputed <- impute(norm, 'MinProb')

# Test for differential enriched proteins
diff <- test_diff(imputed, 'WT', 'control')
final <- add_rejections(diff, alpha = 0.05, lfc = 4.5)

# Merge iBAQ intensities of proteins that have shared peptides
ibaq <- merge_ibaq(data_unique, peptides)

# Calculate relative stoichiometry versus 'Suz12' in the 'GFP_vs_WT' contrast
stoi <- get_stoichiometry(final, ibaq, contrast = 'GFP_vs_WT', bait = 'Suz12')

}
