% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_suffix}
\alias{get_suffix}
\title{Obtain the longest common suffix}
\usage{
get_suffix(words)
}
\arguments{
\item{words}{Character vector,
A list of words.}
}
\value{
A character vector containing the suffix
}
\description{
\code{get_suffix} returns the longest common suffix
of the supplied words.
}
\examples{
# Get suffix
names <- c("xyz_rep", "abc_rep")
get_suffix(names)
}
