% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GFPip}
\alias{GFPip}
\title{GFPip - EED protein interactors (IP-MS dataset)}
\format{A data.frame with 1914 observations and 30 variables:
\describe{
  \item{Protein.IDs}{Uniprot IDs}
  \item{Majority.protein.IDs}{Uniprot IDs of major protein(s) in the protein group}
  \item{Protein.names}{Full protein names}
  \item{Gene.names}{Gene name}
  \item{Fasta.headers}{Header as present in the Uniprot fasta file}
  \item{Peptides}{Number of peptides identified for this protein group}
  \item{Razor...unique.peptides}{Number of peptides used for the quantification of this protein group}
  \item{Unique.peptides}{Number of peptides identified which are unique for this protein group}
  \item{Intensity columns (6)}{Raw mass spectrometry intensity, A.U.}
  \item{iBAQ columns (6)}{iBAQ normalized mass spectrometry intensity, A.U.}
  \item{LFQ.intensity columns (6)}{LFQ normalized mass spectrometry intensity, A.U.}
  \item{Only.identified.by.site}{The protein is only identified by a modification site if marked ('+')}
  \item{Reverse}{The protein is identified in the decoy database if marked ('+')}
  \item{Contaminant}{The protein is a known contaminant if marked ('+')}
  \item{id}{The protein group ID}
}}
\source{
Kloet et al (2016).
The dynamic interactome and genomic targets of
Polycomb complexes during stem-cell differentiation.
Nature Structural & Molecular Biology 23(7): 682-90.
\href{http://www.nature.com/nsmb/journal/v23/n7/full/nsmb.3248.html}{doi: 10.1038/nsmb.3248}
}
\usage{
GFPip
}
\value{
A data.frame with the GFPip dataset
}
\description{
The GFPip dataset contains label free quantification (LFQ)
and intensity-based absolute quantification (iBAQ) data for
EED protein interactors in embryonic stem cells,
generated by Kloet et al 2016.
The dataset contains the proteingroups output
file from \href{http://www.maxquant.org}{MaxQuant}.
}
\keyword{datasets}
