% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GFPip_ExpDesign}
\alias{GFPip_ExpDesign}
\title{Experimental design of the GFPip dataset}
\format{A data.frame with 6 observations and 3 variables:
\describe{
  \item{label}{Label names}
  \item{condition}{Experimental conditions}
  \item{replicate}{Replicate number}
}}
\source{
Kloet et al (2016).
The dynamic interactome and genomic targets of
Polycomb complexes during stem-cell differentiation.
Nature Structural & Molecular Biology 23(7): 682-90.
\href{http://www.nature.com/nsmb/journal/v23/n7/full/nsmb.3248.html}{doi: 10.1038/nsmb.3248}
}
\usage{
GFPip_ExpDesign
}
\value{
A data.frame.
}
\description{
The GFPip_ExpDesign object annotates
6 different samples of the GFPip dataset
in 2 conditions and 3 replicates.
}
\keyword{datasets}
