% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GFPip_pep}
\alias{GFPip_pep}
\title{GFPip_pep - Peptides table of GFP affinity purification of PRC2 in ESCs followed by mass spectrometry}
\format{A data.frame with 12750 observations and 29 variables:
\describe{
  \item{Sequence}{Peptide sequence}
  \item{Mass}{Mass of peptide (Da)}
  \item{Proteins}{Uniprot IDs of proteins to which the peptide is matched}
  \item{Leading.razor.protein}{Uniprot ID of the main protein}
  \item{Gene.names}{Gene names of proteins to which the peptide is matched}
  \item{Unique..Groups.}{Whether the peptide is unique to the protein group to which the peptide is matched}
  \item{Unique..Proteins.}{Whether the peptide is unique to the protein in the complete database}
  \item{Charges}{Peptide charge}
  \item{PEP}{Posterior Error Probability of the identification}
  \item{Score}{Andromeda score}
  \item{Reverse}{The peptide is identified in the decoy database if marked ("+")}
  \item{Contaminant}{The peptide is a known contaminant if marked ("+")}
  \item{id}{The peptide ID}
  \item{Protein.group.IDs}{The protein group ID of the protein to which the peptide is matched}
}}
\source{
Kloet et al (2016). The dynamic interactome and genomic targets of Polycomb complexes during stem-cell differentiation. Nature Structural & Molecular Biology:
  \url{http://www.nature.com/nsmb/journal/v23/n7/full/nsmb.3248.html}
}
\usage{
GFPip_pep
}
\description{
This dataset contains the identified peptides (MaxQuant output \url{http://www.maxquant.org})
}
\keyword{datasets}
