% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{iBAQ}
\alias{iBAQ}
\title{iBAQ workflow}
\usage{
iBAQ(results, peptides, contrast, bait, level = 1)
}
\arguments{
\item{results}{List of SummarizedExperiments
obtained from the \code{\link{LFQ}} wrapper function.}

\item{peptides}{Data.frame,
Peptide table from MaxQuant ('peptides.txt').}

\item{contrast}{Character,
The specific contrast for which to calculate stoichiometries.}

\item{bait}{Character,
The name of the protein to which all other proteins
will be scaled for the relative stoichiometry.}

\item{level}{Numerical,
The level to which the bait will be scaled}
}
\value{
A data.frame with the relative stoichiometry data.
}
\description{
\code{iBAQ} is a wrapper function running the entire analysis workflow
for stoichiometry analysis using
intensity-based absolute quantification (iBAQ)-based proteomics data.
}
\examples{
# load data and test for differentially enriched proteins
data <- GFPip
expdesign <- GFPip_ExpDesign
results <- LFQ(data, expdesign, 'MinProb', 'WT', 'control',
   filter = c('Reverse', 'Contaminant'), alpha = 0.05, lfc = 4.5)

# load peptide data and perform iBAQ-based stoichiometry analysis
peptides <- GFPip_pep
iBAQ(results, peptides, contrast = 'GFP_vs_WT', bait = 'Suz12')
}
