% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{ibaq_merge}
\alias{ibaq_merge}
\title{Merge iBAQ intensities of protein groups based on shared peptides}
\usage{
ibaq_merge(data, peptides)
}
\arguments{
\item{data}{Data.frame, Data object with unique names generated by \code{\link{unique_names}}.}

\item{peptides}{Data.frame, Peptide table from MaxQuant ("peptides.txt").}
}
\value{
A data.frame with iBAQ intensities per protein group from peptides unique to this group
}
\description{
\code{ibaq_merge} generates a data.frame with iBAQ intensities, which are merged for proteins with shared peptides.
}
\examples{
data <- GFPip
peptides <- GFPip_pep

ibaq <- ibaq_merge(data, peptides)
colnames(ibaq)
head(ibaq)
}
