% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{imputation_MSn}
\alias{imputation_MSn}
\title{Impute missing values}
\usage{
imputation_MSn(data, fun)
}
\arguments{
\item{data}{ExpressionSet, Data object for which missing values will be imputed.}

\item{fun}{Character, Function used for data imputation based on \code{\link{impute}}.}
}
\value{
An imputed ExpressionSet object.
}
\description{
\code{imputation_MSn} imputes missing values based on \code{\link{impute}}.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_exprset <- into_exprset_expdesign(example_unique, columns, exp_design)

example_vsn <- norm_vsn(example_exprset)
example_filter <- miss_val_filter(example_vsn, thr = 0)
example_impute_MinProb <- imputation_MSn(example_filter, fun = "MinProb")
example_impute_QRILC <- imputation_MSn(example_filter, fun = "QRILC")
}
