% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{imputation_perseus}
\alias{imputation_perseus}
\title{Impute missing values manually}
\usage{
imputation_perseus(data, shift = 1.8, scale = 0.3)
}
\arguments{
\item{data}{ExpressionSet, Data object for which missing values will be imputed.}

\item{shift}{Integer, sets the left-shift based on the original distribution (1 corresponds to 1 sd left-shift from median).}

\item{scale}{Integer, sets the widht based on the original distribution sd.}
}
\value{
An imputed ExpressionSet object.
}
\description{
\code{imputation_perseus} imputes missing values by a distribution of values which is left-shifted and scaled from the original distribution.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_exprset <- into_exprset_expdesign(example_unique, columns, exp_design)

example_vsn <- norm_vsn(example_exprset)
example_filter <- miss_val_filter(example_vsn, thr = 0)
example_imp <- imputation_perseus(example_filter, shift = 1.8, scale = 0.3)
}
