% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{into_exprset}
\alias{into_exprset}
\title{Data.frame to ExpressionSet}
\usage{
into_exprset(data, columns)
}
\arguments{
\item{data}{Data.frame, The data object which will be turned into an ExpressionSet.}

\item{columns}{Vector of integers, Number of the columns that contain the Expression data.}
}
\value{
An ExpressionSet object.
}
\description{
\code{into_exprset} creates an ExpressionSet object based on a data.frame.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
example_exprset <- into_exprset(example_unique, columns)
}
