% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{into_exprset_expdesign}
\alias{into_exprset_expdesign}
\title{Data.frame to ExpressionSet using an experimental design}
\usage{
into_exprset_expdesign(data, columns, expdesign)
}
\arguments{
\item{data}{Data.frame, Data object which will be turned into an ExpressionSet.}

\item{columns}{Vector of integers, Number of the columns that contain the Expression data.}

\item{expdesign}{Data.frame, Experimental design object}
}
\value{
An ExpressionSet object.
}
\description{
\code{into_exprset_expdesign} creates an ExpressionSet object based on two data.frames: the data and experimental design.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_exprset <- into_exprset_expdesign(example_unique, columns, exp_design)
}
