% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{linear_model}
\alias{linear_model}
\title{Linear model test}
\usage{
linear_model(data, control, type)
}
\arguments{
\item{data}{ExpressionSet, Data object for which the variance will be analyzed.}

\item{control}{Character, The sample name to which the contrasts are generated (the control sample would be most appropriate).}

\item{type}{all" or "control" The type of contrasts that will be generated.}
}
\value{
An ExpressionSet object containing FDR estimates of differential expression.
}
\description{
\code{linear_model} performs a differential expression test based on linear models and empherical Bayes statistics (\code{\link{limma}}).
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
example_exprset <- into_exprset(example_unique, columns)

example_vsn <- norm_vsn(example_exprset)
example_filter <- miss_val_filter(example_vsn, thr = 0)
example_impute <- imputation_MSn(example_filter, fun = "QRILC")

example_lm <- linear_model(example_impute, "Con_", "control")
}
