% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{manual_impute}
\alias{manual_impute}
\title{Imputation by random draws from a manually defined distribution}
\usage{
manual_impute(data, scale = 0.3, shift = 1.8)
}
\arguments{
\item{data}{SummarizedExperiment, Data object for which missing values will be imputed.}

\item{scale}{Integer}

\item{shift}{Integer}
}
\value{
An imputed SummarizedExperiment object.
}
\description{
\code{manual_impute} missing values in a Summarized
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
example_vsn <- norm_vsn(example_filter)

example_impute_manual <- imputation(example_vsn, fun = "man", shift = 1.8, scale = 0.3)
}
