% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{merge_ibaq}
\alias{merge_ibaq}
\title{Merge iBAQ intensities of protein groups based on shared peptides}
\usage{
merge_ibaq(data, peptides)
}
\arguments{
\item{data}{Data.frame, Data object with unique names generated by \code{\link{unique_names}}.}

\item{peptides}{Data.frame, Peptide table from MaxQuant ("peptides.txt").}
}
\value{
A data.frame with iBAQ intensities per protein group from peptides unique to this group
}
\description{
\code{merge_ibaq} generates a data.frame with iBAQ intensities, which are merged for proteins with shared peptides.
}
\examples{
data <- GFPip
data_unique <- unique_names(data, "Gene.names", "Protein.IDs")
peptides <- GFPip_pep

ibaq <- merge_ibaq(data_unique, peptides)
colnames(ibaq)
head(ibaq)
}
