% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{miss_val_filter}
\alias{miss_val_filter}
\title{Filter on missing values}
\usage{
miss_val_filter(data, thr = 0)
}
\arguments{
\item{data}{ExpressionSet, Data object which will be filtered.}

\item{thr}{Integer, sets the threshold for the allowed number of missing values per condition.}
}
\value{
An filtered ExpressionSet object.
}
\description{
\code{miss_val_filter} filters an ExpressionSet object based on missing values.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_exprset <- into_exprset_expdesign(example_unique, columns, exp_design)

example_vsn <- norm_vsn(example_exprset)
example_stringent_filter <- miss_val_filter(example_vsn, thr = 0)
example_less_stringent_filter <- miss_val_filter(example_vsn, thr = 1)
}
