% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{msn2se}
\alias{msn2se}
\title{MSnSet to SummarizedExperiment object conversion}
\usage{
msn2se(msn)
}
\arguments{
\item{msn}{MSnSet object,
Object which will be turned into a SummarizedExperiment object.}
}
\value{
A SummarizedExperiment object.
}
\description{
\code{msn2se} generates a SummarizedExperiment object from a MSnSet object.
}
\examples{
# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Convert to MSnSet
data_msn <- se2msn(se)
# Convert back to SE
data_se <- msn2se(data_msn)
}
