% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{norm_vsn}
\alias{norm_vsn}
\title{Data normalization using vsn}
\usage{
norm_vsn(data)
}
\arguments{
\item{data}{ExpressionSet, Data object which will be normalized.}
}
\value{
An normalized ExpressionSet object.
}
\description{
\code{norm_vsn} nomralizes an ExpressionSet object using \code{\link{vsn}}.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
example_exprset <- into_exprset(example_unique, columns)

example_vsn <- norm_vsn(example_exprset)
}
