% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalize}
\alias{normalize}
\title{Normalization using vsn}
\usage{
normalize(data)
}
\arguments{
\item{data}{SummarizedExperiment, Proteomics data with log2-transformed values (as data obtained from \code{\link{make_se}}.}
}
\value{
A normalized SummarizedExperiment object.
}
\description{
\code{normalize} performs variance stabilization on a SummarizedExperiment object using \code{\link[vsn]{vsn-package}}.
}
\examples{
data <- UbiLength
data <- filter(data, Reverse != "+", Potential.contaminant != "+")
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize(filt)
}
