% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_aggregation.R
\name{normalize_pe}
\alias{normalize_pe}
\title{Normalize a QFeatures object}
\usage{
normalize_pe(
  pe,
  method = c("diff.median", "quantiles", "quantiles.robust", "vsn"),
  i = "peptideImp",
  name = "peptideNorm"
)
}
\arguments{
\item{pe}{A QFeature object}

\item{method}{Character(1), normalisation method, one of "diff.median", "quantiles", "quantiles.robust" or "vsn".
See \link[QFeatures]{normalize}.}

\item{i}{A numeric or a character giving the index or the name, respectively, of the assay(s) to be processed.}

\item{name}{Character(1) naming the new normalized assay.}
}
\description{
Normalize the \code{i} assay in QFeatures object though \link[QFeatures:QFeatures-processing]{QFeatures::normalize} function.
The normalized assay is saved in the \code{name} assay.
}
\examples{
\dontrun{
# Load example peptide data
data(Silicosis_peptide)
ecols <- grep("Intensity.", colnames(Silicosis_peptide), value = TRUE)

# Construct QFeature object
pe_peptides <- make_pe_parse(Silicosis_peptide, columns = ecols, remove_prefix = TRUE, log2transform = TRUE,mode = "delim")

# Filter, imputation
pe <- filter_pe(pe_peptides, thr = 1,fraction = 0.4, filter_formula = ~ Reverse != '+' & Potential.contaminant !="+" )
pe <- impute_pe(pe, fun = "QRILC", name = "peptideImp")

# Normalization
pe <- normalize_pe(pe,method = "quantiles", i = "peptideImp", name = "peptideNorm")
}
}
