% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_detect}
\alias{plot_detect}
\title{Visualize intensities of proteins with missing values}
\usage{
plot_detect(data)
}
\arguments{
\item{data}{SummarizedExperiment, Data object with missing values}
}
\value{
Density and CumSum plots of intensities of protein with missing values (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_detect} generates density and CumSum plots of protein intensities with and without missing values
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)

plot_detect(example_filter)
}
