% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot protein overlap between conditions}
\usage{
plot_frequency(data)
}
\arguments{
\item{data}{ExpressionSet, Data object for which to plot observation frequency.}
}
\value{
Barplot of overlap of protein identifications between conditions (generated by \code{\link{ggplot}})
}
\description{
\code{plot_frequency} generates a barplot of the protein overlap between conditions.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_exprset <- make_se(example_unique, columns, exp_design)
plot_frequency(example_exprset)

example_filter <- filter_missval(example_exprset, thr = 0)
plot_frequency(example_filter)
}
