% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{plot_ibaq}
\alias{plot_ibaq}
\title{Plot the iBAQ values versus log fold change}
\usage{
plot_ibaq(data, contrast, labelsize = 3)
}
\arguments{
\item{data}{SummarizedExperiment object, Data object which has been generated by generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}

\item{contrast}{Character, The specific contrast to plot.}

\item{labelsize}{Integer, sets the size of name labels.}
}
\value{
A scatter plot
}
\description{
\code{plot_ibaq} plots a scatter plot of the iBAQ intensities versus the LFQ fold changes.
}
\examples{
# load data
data <- GFPip
exp_design <- GFPip_ExpDesign
peptides <- GFPip_pep

# Make unique names and filter for reverse and contaminant hits
data_unique <- data \%>\% unique_names(., "Gene.names", "Protein.IDs")
data_filt <- data_unique \%>\% filter(Reverse != "+", Contaminant != "+")

# Make a SummarizedExperiment and perform filtering on missing values,
# variance stabilization and imputation on this object
cols <- grep("LFQ", colnames(data_filt))
se <- make_se(data_filt, cols, exp_design)
filt <- filter_missval(se)
norm <- norm_vsn(filt)
imp <- imputation(norm, "MinProb")

# Test for differential enriched proteins
lm <- linear_model(imp, "WT", "control")
sign <- cutoffs(lm, alpha = 0.05, lfc = 4.5)

# Make a results table and check number of differential enrich proteins
results <- results(sign)
results \%>\% filter(sign == "+") \%>\% nrow()

# Plot iBAQ vs LFQ plot
plot_ibaq(sign, "GFP_vs_WT", labelsize = 3)

}
