% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_imp}
\alias{plot_imp}
\title{Visualize imputation}
\usage{
plot_imp(raw, imp)
}
\arguments{
\item{raw}{SummarizedExperiment, Data object before imputation}

\item{imp}{SummarizedExperiment, Data object after imputation by \code{\link{imputation}}.}
}
\value{
Density plots for all conditions before and after imputated (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_imp} generates density plots for all conditions before and after imputation
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
example_vsn <- norm_vsn(example_filter)
example_imp <- imputation(example_vsn, "MinProb")

plot_imp(example_vsn, example_imp)
}
