% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_missval}
\alias{plot_missval}
\title{Plot a heatmap of proteins with missing values}
\usage{
plot_missval(data)
}
\arguments{
\item{data}{SummarizedExperiment,
Data object which has been generated by
\code{\link{test_diff}} and \code{\link{add_rejections}}.}
}
\value{
A heatmap indicating whether values are missing (0) or not (1)
(generated by \code{\link[ComplexHeatmap]{Heatmap}})
}
\description{
\code{plot_missval} generates a heatmap of proteins
with missing values to discover whether values are missing by random or not.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != '+' & data$Potential.contaminant != '+',]
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs', delim = ';')

columns <- grep('LFQ.', colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
plot_missval(filt)
}
