% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_norm}
\alias{plot_norm}
\title{Visualize normalization}
\usage{
plot_norm(raw, norm)
}
\arguments{
\item{raw}{SummerizedExperiment, Data object before normalization.}

\item{norm}{SummerizedExperiment, Data object after normalization by \code{\link{norm_vsn}}.}
}
\value{
Boxplots for all conditions before and after normalization (generated by \code{\link{ggplot}})
}
\description{
\code{plot_norm} generates boxplots for all conditions before and after normalization.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
example_vsn <- norm_vsn(example_filter)

plot_norm(example_filter, example_vsn)
}
