% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_numbers}
\alias{plot_numbers}
\title{Plot protein numbers}
\usage{
plot_numbers(data)
}
\arguments{
\item{data}{SummerizedExperiment, Data object for which to plot protein numbers.}
}
\value{
Barplot of the number of identified proteins per sample (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_numbers} generates a barplot of the number of identified proteins per sample.
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)
plot_numbers(example_se)

example_filter <- filter_missval(example_se, thr = 0)
plot_numbers(example_filter)
}
