% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_single}
\alias{plot_single}
\title{Plot values for a protein of interest}
\usage{
plot_single(data, protein, type)
}
\arguments{
\item{data}{SummarizedExperiment, Data object which has been generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}

\item{protein}{Character, Protein name.}

\item{type}{"centered" or "contrast", The type of data scaling used for plotting.}
}
\value{
A barplot (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_single} generates a barplot of a protein of interest.
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
example_vsn <- norm_vsn(example_filter)
example_impute <- imputation(example_filter, fun = "MinProb", q = 0.01)

example_lm <- linear_model(example_impute, "Ctrl", "control")
example_sign <- cutoffs(example_lm, alpha = 0.05, lfc = 1)

plot_single(example_sign, "USP15", "centered")
plot_single(example_sign, "USP15", "contrast")
}
