% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{plot_stoi}
\alias{plot_stoi}
\title{Plot the relative stoichiometry}
\usage{
plot_stoi(data, thr = 0.01, max_y = NULL)
}
\arguments{
\item{data}{Data.frame, Stoichiometry object generated by \code{\link{stoichiometry}}.}

\item{thr}{Numerical, The stoichiometry threshold above which proteins will be plotted.}

\item{max_y}{Numerical, The maximum scale of the y-axis.}
}
\value{
A barplot
}
\description{
\code{plot_stoi} plots a barplot of the relative stoichiometries of all proteins.
}
\examples{
#' # load data
data <- GFPip
exp_design <- GFPip_ExpDesign
peptides <- GFPip_pep

# Make unique names and filter for reverse and contaminant hits
data_unique <- data \%>\% unique_names(., "Gene.names", "Protein.IDs")
data_filt <- data_unique \%>\% filter(Reverse != "+", Contaminant != "+")

# Make a SummarizedExperiment and perform filtering on missing values, variance stabilization and imputation on this object
cols <- grep("LFQ", colnames(data_filt))
se <- make_se(data_filt, cols, exp_design)
filt <- filter_missval(se)
norm <- norm_vsn(filt)
imp <- imputation(norm, "MinProb")

# Test for differential enriched proteins
lm <- linear_model(imp, "WT", "control")
sign <- cutoffs(lm, alpha = 0.05, lfc = 4.5)

# Make a results table and check number of differential enrich proteins
results <- results(sign)
results \%>\% filter(sign == "+") \%>\% nrow()

# Merge iBAQ intensities of proteins that have shared peptides
ibaq <- ibaq_merge(data_unique, peptides)

# Calculate relative stoichiometry versus "Suz12" in the "GFP_vs_WT" contrast
stoi <- stoichiometry(sign, ibaq, contrast = "GFP_vs_WT", bait = "Suz12")
plot_stoi(stoi)

}
