% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ_functions.R
\name{plot_stoichiometry}
\alias{plot_stoichiometry}
\title{Plot relative stoichiometries}
\usage{
plot_stoichiometry(data, thr = 0.01, max_y = NULL)
}
\arguments{
\item{data}{Data.frame,
Stoichiometry table generated by \code{\link{get_stoichiometry}}.}

\item{thr}{Numerical,
The stoichiometry threshold above which proteins will be plotted.}

\item{max_y}{Numerical,
The maximum scale of the y-axis.}
}
\value{
A barplot (generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{plot_stoichiometry} plots a barplot of the
relative stoichiometries.
}
\examples{
# load data
data <- GFPip
exp_design <- GFPip_ExpDesign
peptides <- GFPip_pep

# Make unique names and filter for reverse and contaminant hits
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs')
data_filt <- data_unique[data_unique$Reverse != '+' & data_unique$Contaminant != '+',]

# Make a SummarizedExperiment and perform filtering on missing values,
# variance stabilization and imputation on this object
cols <- grep('LFQ', colnames(data_filt))
se <- make_se(data_filt, cols, exp_design)
filt <- filter_missval(se)
norm <- normalize_vsn(filt)
imputed <- impute(norm, 'MinProb')

# Test for differential enriched proteins
diff <- test_diff(imputed, 'WT', 'control')
final <- add_rejections(diff, alpha = 0.05, lfc = 4.5)

# Merge iBAQ intensities of proteins that have shared peptides
ibaq <- merge_ibaq(data_unique, peptides)

# Calculate relative stoichiometry versus 'Suz12' in the 'GFP_vs_WT' contrast
stoi <- get_stoichiometry(final, ibaq, contrast = 'GFP_vs_WT', bait = 'Suz12')
plot_stoichiometry(stoi)

}
