% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(data, contrast, labelsize = 3, add_names = TRUE)
}
\arguments{
\item{data}{SummarizedExperiment, Data object which has been generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}

\item{contrast}{Character, The specific contrast to plot.}

\item{labelsize}{Integer, sets the size of name labels.}

\item{add_names}{Boolean, whether or not to plot names.}
}
\value{
A volcano plot (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_volcano} generates a volcano plot for a specified contrast.
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
example_vsn <- norm_vsn(example_filter)
example_impute <- imputation(example_filter, fun = "MinProb", q = 0.01)

example_lm <- linear_model(example_impute, "Ctrl", "control")
example_sign <- cutoffs(example_lm, alpha = 0.05, lfc = 1)

plot_volcano(example_sign, "Ubi6_vs_Ctrl", labelsize = 5, add_names = TRUE)
plot_volcano(example_sign, "Ubi6_vs_Ctrl", add_names = FALSE)
plot_volcano(example_sign, "Ubi4_vs_Ctrl", labelsize = 5, add_names = TRUE)
plot_volcano(example_sign, "Ubi4_vs_Ctrl", add_names = FALSE)
}
