% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{results_stoi}
\alias{results_stoi}
\title{Relative stoichiometry results table}
\usage{
results_stoi(data, thr = 0.01)
}
\arguments{
\item{data}{Data.frame, Stoichiometry object generated by \code{\link{stoichiometry}}.}

\item{thr}{Numerical, The stoichiometry threshold above which proteins will be plotted.}
}
\value{
A data.frame with the stoichiometry information
}
\description{
\code{results_stoi} generates a results table of relative stoichiometry values.
}
\examples{
# load data
data <- GFPip
exp_design <- GFPip_ExpDesign
peptides <- GFPip_pep

# Make unique names and filter for reverse and contaminant hits
data_unique <- data \%>\% unique_names(., "Gene.names", "Protein.IDs")
data_filt <- data_unique \%>\% filter(Reverse != "+", Contaminant != "+")

# Make a SummarizedExperiment and perform filtering on missing values,
# variance stabilization and imputation on this object
cols <- grep("LFQ", colnames(data_filt))
se <- make_se(data_filt, cols, exp_design)
filt <- filter_missval(se)
norm <- norm_vsn(filt)
imp <- imputation(norm, "MinProb")

# Test for differential enriched proteins
lm <- linear_model(imp, "WT", "control")
sign <- cutoffs(lm, alpha = 0.05, lfc = 4.5)

# Make a results table and check number of differential enrich proteins
results <- results(sign)
results \%>\% filter(sign == "+") \%>\% nrow()

# Merge iBAQ intensities of proteins that have shared peptides
ibaq <- ibaq_merge(data_unique, peptides)

# Calculate relative stoichiometry versus "Suz12" in the "GFP_vs_WT" contrast
stoi <- stoichiometry(sign, ibaq, contrast = "GFP_vs_WT", bait = "Suz12")
plot_stoi(stoi)
results_stoi(stoi)

}
