% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app_function.R
\name{run_app}
\alias{run_app}
\title{proteomeR shiny apps}
\usage{
run_app(app)
}
\arguments{
\item{app}{The name of the app. Currently supporting "LFQ" or "TMT".}
}
\description{
\code{run_app} Launches an interactive shiny app for interactive analysis of proteomics data.
}
\examples{
\dontrun{
# Run the app
run_app("LFQ")

run_app("TMT")

}
}
