% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{single_prot_plot}
\alias{single_prot_plot}
\title{Plot values for a protein of interest}
\usage{
single_prot_plot(data, protein, type)
}
\arguments{
\item{data}{ExpressionSet, Data object which has been generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}

\item{protein}{Character, Protein name.}

\item{type}{"centered" or "contrast", The type of data scaling used for plotting.}
}
\value{
A barplot generated by \code{\link{ggplots2}}
}
\description{
\code{single_prot_plot} generates a barplot of a protein of interest.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
example_exprset <- into_exprset(example_unique, columns)
example_vsn <- norm_vsn(example_exprset)
example_filter <- miss_val_filter(example_vsn, thr = 0)
example_impute <- imputation_MSn(example_filter, fun = "QRILC")

example_lm <- linear_model(example_impute, "Con_", "control")
example_sign <- cutoffs(example_lm, alpha = 0.05, lfc = 1)

single_prot_plot(example_sign, "USP15", "centered")
single_prot_plot(example_sign, "USP15", "contrast")
}
