% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{stoichiometry}
\alias{stoichiometry}
\title{Relative stoichiometry}
\usage{
stoichiometry(data, ibaq, contrast, bait, level = 1)
}
\arguments{
\item{data}{SummerizedExperiment, Data object generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}

\item{ibaq}{Data.frame, iBAQ object generated by \code{\link{ibaq_merge}}.}

\item{contrast}{Character, The specific contrast to calculate the stroichiometry for.}

\item{bait}{Character, The name of the protein to which all other proteins will be scaled for the relative stoichiometry}

\item{level}{Numerical, The level to which the bait will be scaled}
}
\value{
A data.frame with relative stoichiometry values per protein
}
\description{
\code{stoichiometry} calculates relative stoichiometry of proteins
}
\examples{
# load data
data <- GFPip
exp_design <- GFPip_ExpDesign
peptides <- GFPip_pep

# Make unique names and filter for reverse and contaminant hits
data_unique <- data \%>\% unique_names(., "Gene.names", "Protein.IDs")
data_filt <- data_unique \%>\% filter(Reverse != "+", Contaminant != "+")

# Make a SummarizedExperiment and perform filtering on missing values,
# variance stabilization and imputation on this object
cols <- grep("LFQ", colnames(data_filt))
se <- make_se(data_filt, cols, exp_design)
filt <- filter_missval(se)
norm <- norm_vsn(filt)
imp <- imputation(norm, "MinProb")

# Test for differential enriched proteins
lm <- linear_model(imp, "WT", "control")
sign <- cutoffs(lm, alpha = 0.05, lfc = 4.5)

# Make a results table and check number of differential enrich proteins
results <- results(sign)
results \%>\% filter(sign == "+") \%>\% nrow()

# Merge iBAQ intensities of proteins that have shared peptides
ibaq <- ibaq_merge(data_unique, peptides)

# Calculate relative stoichiometry versus "Suz12" in the "GFP_vs_WT" contrast
stoi <- stoichiometry(sign, ibaq, contrast = "GFP_vs_WT", bait = "Suz12")

}
