% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_theme.R
\name{theme_DEP2}
\alias{theme_DEP2}
\title{DEP ggplot theme 2}
\usage{
theme_DEP2()
}
\description{
\code{theme_DEP2} is the ggplot theme's used for plotting in \code{\link{DEP}} with vertical x-axis labels.
}
