% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unique_names}
\alias{unique_names}
\title{Unique names}
\usage{
unique_names(data, name, ids, delim = ";")
}
\arguments{
\item{data}{Data.frame, The data object for which unique names will be created.}

\item{name}{Character, Name of the column containing feature names.}

\item{ids}{Character, Name of the column containing feature IDs.}

\item{delim}{Character, Delimiter seperating the proteins within on protein group.}
}
\value{
A data.frame containing an addiontal variable "name" that contains unique names.
}
\description{
\code{unique_names} generates unique identifiers for a dataset based on a "name" and "id" column.
}
\examples{
example <- UbiLength
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")
}
