% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cutoffs}
\alias{cutoffs}
\title{Denote significant proteins}
\usage{
cutoffs(data, alpha = 0.05, lfc = 1)
}
\arguments{
\item{data}{SummerizedExperiment, Data object which will be filtered for significant proteins.}

\item{alpha}{Integer, sets the threshold for the false discovery rate (FDR).}

\item{lfc}{Integer, sets the threshold for the log fold change (lfc).}
}
\value{
An SummerizedExperiment object containing a variable "sign" denoting the significant proteins.
}
\description{
\code{cutoffs} denotes significant proteins based on defined cutoffs.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
example_vsn <- norm_vsn(example_filter)
example_impute <- imputation(example_filter, fun = "MinProb", q = 0.01)

example_lm <- linear_model(example_impute, "Ctrl", "control")
example_sign <- cutoffs(example_lm, alpha = 0.05, lfc = 1)
significant_proteins <- example_sign[rowData(example_sign)$sign == "+", ]
nrow(significant_proteins)
}
