% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP.R
\docType{package}
\name{DEP}
\alias{DEP}
\alias{DEP-package}
\title{DEP: A package for Differential Enrichment analysis of Proteomics data.}
\description{
The DEP package provides all functionalities to analyze label free protein quantification (LFQ) and tandem-mass-tags (TMT) labeled proteomics data
}
\section{Shiny apps}{

\itemize{
  \item \code{\link{run_app}}: Shiny apps for interactive analysis
}
}

\section{Workflows}{

\itemize{
  \item \code{\link{LFQ}}: Label-free quantification (LFQ) workflow
  \item \code{\link{TMT}}: Tandem-mass-tags (TMT) workflow
  \item \code{\link{report}}: Create a rmarkdown report
  \item \code{\link{iBAQ}}: iBAQ-based stoichiometry workflow
}
}

\section{Main functions}{

\itemize{
  \item \code{\link{make_unique}}: Generate unique names
  \item \code{\link{make_se_parse}}: Turn data.frame into SummarizedExperiment by parsing column names
  \item \code{\link{make_se}}: Turn data.frame into SummarizedExperiment using an experimental design
  \item \code{\link{filter_missval}}: Filter on missing values
  \item \code{\link{normalize_vsn}}: Normalize data using vsn
  \item \code{\link{impute}}: Impute missing values
  \item \code{\link{test_diff}}: Differential enrichment analysis
  \item \code{\link{add_rejections}}: Mark significant proteins
  \item \code{\link{get_results}}: Generate a results table
}
}

\section{Visualization functions}{

\itemize{
  \item \code{\link{plot_single}}: Barplot for a protein of interest
  \item \code{\link{plot_volcano}}: Volcano plot for a specified contrast
  \item \code{\link{plot_heatmap}}: Heatmap of all significant proteins
  \item \code{\link{plot_ibaq}}: Scatter plot of iBAQ intensities versus log fold changes for a specified contrast
  \item \code{\link{plot_normalization}}: Boxplots to inspect normalization
  \item \code{\link{plot_detect}}: Density and CumSum plots to inspect detection limit
  \item \code{\link{plot_imputation}}: Density plots to inspect imputation
  \item \code{\link{plot_missval}}: Heatmap to inspect missing values
  \item \code{\link{plot_numbers}}: Barplot of proteins identified
  \item \code{\link{plot_frequency}}: Barplot of protein identification overlap between conditions
  \item \code{\link{plot_coverage}}: Barplot of the protein coverage in conditions
  \item \code{\link{plot_pca}}: PCA plot of top variable proteins
  \item \code{\link{plot_corr}}: Correlation matrix
  \item \code{\link{plot_cond_freq}}: Barplot of number of significant conditions per protein
  \item \code{\link{plot_cond_overlap}}: Barplot of number of proteins per condition(s)
}
}

\section{iBAQ functions}{

\itemize{
  \item \code{\link{merge_ibaq}}: Merge iBAQ intensities of protein groups with shared peptides
  \item \code{\link{get_stoichiometry}}: Calculate relative stoichiometry of all significant proteins
  \item \code{\link{plot_stoichiometry}}: Barplot of relative stoichiometries
}
}

\section{Additional functions}{

\itemize{
  \item \code{\link{se2msn}}: SummarizedExperiment to MSnSet object conversion
  \item \code{\link{manual_impute}}: Imputation by random draws from a manually defined distribution
  \item \code{\link{get_prefix}}: Obtain the longest common prefix
}
}

\section{Example data}{

\itemize{
  \item \code{\link{UbiLength}}: Ubiquitin interactors of different linear ubiquitin lengths (UbIA-MS dataset) (Zhang, Smits, van Tilburg et al. Mol. Cell 2017)
  \item \code{\link{UbiLength_ExpDesign}}: Experimental design of the UbiLength dataset
  \item \code{\link{DiUbi}}: Ubiquitin interactors for different diubiquitin-linkages (UbIA-MS dataset) (Zhang, Smits, van Tilburg et al. Mol. Cell 2017)
  \item \code{\link{DiUbi_ExpDesign}}: Experimental design of the DiUbi dataset
  \item \code{\link{GFPip}}: EED interactors in ESCs (GFP IP - MS dataset) (Kloet et al. Nat. Struct. Mol. Biol. 2016)
  \item \code{\link{GFPip_ExpDesign}}: Experimental design of the GFPip dataset
  \item \code{\link{GFPip_pep}}: Peptides table of the GFPip dataset
}
}

