% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP.R
\docType{package}
\name{DEP}
\alias{DEP}
\alias{DEP-package}
\title{DEP: A package for Differential Enrichment analysis of Proteomics data.}
\description{
The DEP package provides all functionalities to analyze label free protein quantification (LFQ) and tandem-mass-tags (TMT) labeled proteomics data
}
\section{DEP shiny apps}{

\itemize{
  \item \code{\link{run_app}}: Shiny apps for interactive analysis
}
}

\section{DEP workflows}{

\itemize{
  \item \code{\link{LFQ}}: Label-free quantification (LFQ) workflow
  \item \code{\link{TMT}}: Tandem-mass-tags (TMT) workflow
  \item \code{\link{report}}: Create a rmarkdown report
  \item \code{\link{iBAQ}}: iBAQ-based stoichiometry workflow
}
}

\section{DEP functions}{

\itemize{
  \item \code{\link{unique_names}}: Generate unique names
  \item \code{\link{make_se_parse}}: Turn data.frame into SummarizedExperiment by parsing column names
  \item \code{\link{make_se}}: Turn data.frame into SummarizedExperiment using an experimental design
  \item \code{\link{filter_missval}}: Filter on missing values
  \item \code{\link{norm_vsn}}: Normalize data using vsn
  \item \code{\link{imputation}}: Impute missing values
  \item \code{\link{linear_model}}: Differential expression analysis (limma)
  \item \code{\link{anova_tukey}}: Differential expression analysis (anova - tukey)
  \item \code{\link{cutoffs}}: Denote significant proteins
  \item \code{\link{results}}: Generate a results table
}
}

\section{DEP visualization functions}{

\itemize{
  \item \code{\link{plot_single}}: Barplot for a protein of interest
  \item \code{\link{plot_volcano}}: Volcano plot for a specified contrast
  \item \code{\link{plot_heatmap}}: Heatmap of all significant proteins
  \item \code{\link{plot_norm}}: Boxplots to inspect normalization
  \item \code{\link{plot_detect}}: Density and CumSum plots to inspect detection limit
  \item \code{\link{plot_imp}}: Density plots to inspect imputation
  \item \code{\link{plot_missval}}: Heatmap to inspect missing values
  \item \code{\link{plot_numbers}}: Barplot of proteins identified
  \item \code{\link{plot_frequency}}: Barplot of protein identification overlap between conditions
  \item \code{\link{plot_coverage}}: Barplot of the protein coverage in conditions
}
}

\section{DEP iBAQ functions}{

\itemize{
  \item \code{\link{ibaq_merge}}: Merge iBAQ intensities of protein groups based on shared peptides
  \item \code{\link{stoichiometry}}: Calculate relative stoichiometry of all significant proteins
  \item \code{\link{plot_stoi}}: Barplot of the relative stoichiometries
}
}

