% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GFPip}
\alias{GFPip}
\title{GFPip - EED interactors in ESCs (GFP IP - MS dataset)}
\format{A data.frame with 1914 observations and 30 variables:
\describe{
  \item{Protein.IDs}{Uniprot IDs}
  \item{Majority.protein.IDs}{Uniprot IDs of major protein(s) in the protein group}
  \item{Protein.names}{Full protein names}
  \item{Gene.names}{Gene name}
  \item{Fasta.headers}{Header as present in the Uniprot fasta file}
  \item{Peptides}{Number of peptides identified for this protein group}
  \item{Razor...unique.peptides}{Number of peptides used for the quantification of this protein group}
  \item{Unique.peptides}{Number of peptides identified which are unique for this protein group}
  \item{Intensity columns (6)}{Raw mass spectrometry intensity, A.U.}
  \item{iBAQ columns (6)}{iBAQ normalized mass spectrometry intensity, A.U.}
  \item{LFQ.intensity columns (6)}{LFQ normalized mass spectrometry intensity, A.U.}
  \item{Only.identified.by.site}{The protein is only identified by a modification site if marked ('+')}
  \item{Reverse}{The protein is identified in the decoy database if marked ('+')}
  \item{Contaminant}{The protein is a known contaminant if marked ('+')}
  \item{id}{The protein group ID}
}}
\source{
Kloet et al (2016). The dynamic interactome and genomic targets of Polycomb complexes during stem-cell differentiation. Nature Structural & Molecular Biology:
  \url{http://www.nature.com/nsmb/journal/v23/n7/full/nsmb.3248.html}
}
\usage{
GFPip
}
\value{
A data.frame with the GFPip dataset
}
\description{
This dataset contains label free quantification (LFQ) and intensity-based absolute quantification (iBAQ) data (MaxQuant output \url{http://www.maxquant.org}).
}
\keyword{datasets}
