% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GFPip_ExpDesign}
\alias{GFPip_ExpDesign}
\title{Experimental design of the GFPip dataset}
\format{A data.frame with 6 observations and 3 variables:
\describe{
  \item{label}{Label names}
  \item{condition}{Experimental conditions}
  \item{replicate}{Replicate number}
}}
\source{
Kloet et al (2016). The dynamic interactome and genomic targets of Polycomb complexes during stem-cell differentiation. Nature Structural & Molecular Biology:
  \url{http://www.nature.com/nsmb/journal/v23/n7/full/nsmb.3248.html}
}
\usage{
GFPip_ExpDesign
}
\description{
This dataset annotates 6 different samples in 2 conditions and 3 replicates.
}
\keyword{datasets}
