% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GFPip_pep}
\alias{GFPip_pep}
\title{GFPip_pep - Peptides table of the GFPip dataset}
\format{A data.frame with 12750 observations and 29 variables:
\describe{
  \item{Sequence}{Peptide sequence}
  \item{Mass}{Mass of peptide (Da)}
  \item{Proteins}{Uniprot IDs of proteins to which the peptide is matched}
  \item{Leading.razor.protein}{Uniprot ID of the main protein}
  \item{Gene.names}{Gene names of proteins to which the peptide is matched}
  \item{Unique..Groups.}{Whether the peptide is unique to the protein group to which the peptide is matched}
  \item{Unique..Proteins.}{Whether the peptide is unique to the protein in the complete database}
  \item{Charges}{Peptide charge}
  \item{PEP}{Posterior Error Probability of the identification}
  \item{Score}{Andromeda score}
  \item{Reverse}{The peptide is identified in the decoy database if marked ('+')}
  \item{Contaminant}{The peptide is a known contaminant if marked ('+')}
  \item{id}{The peptide ID}
  \item{Protein.group.IDs}{The protein group ID of the protein to which the peptide is matched}
}}
\source{
Kloet et al (2016).
The dynamic interactome and genomic targets of
Polycomb complexes during stem-cell differentiation.
Nature Structural & Molecular Biology 23(7): 682-90.
\href{http://www.nature.com/nsmb/journal/v23/n7/full/nsmb.3248.html}{doi: 10.1038/nsmb.3248}
}
\usage{
GFPip_pep
}
\value{
A data.frame.
}
\description{
The GFPip_pep object contains the identified peptides
of the GFPip dataset. The dataset contains the peptides table from
\href{http://www.maxquant.org}{MaxQuant}
}
\keyword{datasets}
