% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{impute}
\alias{impute}
\title{Impute missing values}
\usage{
impute(data, fun, ...)
}
\arguments{
\item{data}{SummarizedExperiment,
Proteomics data for which missing values will be imputed.}

\item{fun}{"man", "bpca", "knn", "QRILC", "MLE", "MinDet",
"MinProb", "min", "zero", "mixed" or "nbavg",
Function used for data imputation based on \code{\link{manual_impute}}
and \code{\link[MSnbase]{impute}}.}

\item{...}{Additional arguments for imputation functions as depicted in
\code{\link{manual_impute}} \code{\link[MSnbase]{impute}}.}
}
\value{
An imputed SummarizedExperiment object.
}
\description{
\code{impute} imputes missing values in a proteomics dataset.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)

imputed_MinProb <- impute(norm, fun = "MinProb", q = 0.05)
imputed_QRILC <- impute(norm, fun = "QRILC")

imputed_knn <- impute(norm, fun = "knn", k = 10, rowmax = 0.9)
imputed_MLE <- impute(norm, fun = "MLE")

imputed_manual <- impute(norm, fun = "man", shift = 1.8, scale = 0.3)
}
