% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{LFQ}
\alias{LFQ}
\title{LFQ workflow}
\usage{
LFQ(data, fun, control, type, alpha = 0.05, lfc = 1)
}
\arguments{
\item{data}{Data.frame, The data object.}

\item{fun}{Character, Function used for data imputation based on \code{\link{impute}}.}

\item{control}{Character, The sample name to which the contrasts are generated (the control sample would be most appropriate).}

\item{type}{"all" or "control" The type of contrasts that will be generated.}

\item{alpha}{Numeric, sets the false discovery rate threshold.}

\item{lfc}{Numeric, sets the log fold change threshold.}
}
\value{
A list of two data.frames: 1) \code{results} object containing the significant proteins, 2) \code{data} object containing the full dataset. Additionally, a \code{\link{rmarkdown}} report is generated and saved.
}
\description{
\code{LFQ} is a wrapper function running the entire analysis workflow for label free quantification (LFQ)-based proteomics data.
}
\examples{
example <- UbIA_MS
example_results <- LFQ(example, "QRILC", "Con_", "control", alpha = 0.05, lfc = 1)
}
