% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{LFQ}
\alias{LFQ}
\title{LFQ workflow}
\usage{
LFQ(data, expdesign, fun, control, type, test = NULL, filter = c("Reverse",
  "Potential.contaminant"), name = "Gene.names", ids = "Protein.IDs",
  alpha = 0.05, lfc = 1)
}
\arguments{
\item{data}{Data.frame,
The data object.}

\item{expdesign}{Data.frame,
The experimental design object.}

\item{fun}{Character,
Function used for data imputation based on \code{\link[MSnbase]{impute}}.}

\item{control}{Character,
The sample name to which the contrasts are generated
(the control sample would be most appropriate).}

\item{type}{'all' or 'control',
The type of contrasts that will be generated.}

\item{test}{Character,
The contrasts that will be tested if type = "manual".
These should be formatted as "SampleA_vs_SampleB" or
c("SampleA_vs_SampleC", "SampleB_vs_SampleC").}

\item{filter}{Character,
Name(s) of the column(s) to be filtered on.}

\item{name}{Character,
Name of the column representing gene names.}

\item{ids}{'Character,
Name of the column representing protein IDs.}

\item{alpha}{Numeric,
sets the false discovery rate threshold.}

\item{lfc}{Numeric,
sets the log fold change threshold.}
}
\value{
A list of two data.frames:
1) \code{results} object containing the significant proteins,
2) \code{data} object containing the full dataset.
}
\description{
\code{LFQ} is a wrapper function running the entire analysis workflow
for label free quantification (LFQ)-based proteomics data.
The protein result table from
\href{http://www.nature.com/nbt/journal/v26/n12/full/nbt.1511.html}{MaxQuant}
is used as direct input.
}
\examples{

data <- UbiLength
expdesign <- UbiLength_ExpDesign
results <- LFQ(data, expdesign, 'MinProb', 'Ctrl', 'control')

}
