% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{linear_model}
\alias{linear_model}
\title{Linear model test}
\usage{
linear_model(data, control, type)
}
\arguments{
\item{data}{SummerizedExperiment, Data object for which the variance will be analyzed.}

\item{control}{Character, The condition to which the contrasts are generated (the control would be most appropriate).}

\item{type}{all" or "control" The type of contrasts that will be generated.}
}
\value{
An SummerizedExperiment object containing FDR estimates of differential expression.
}
\description{
\code{linear_model} performs a differential expression test based on linear models and empherical Bayes statistics (\code{\link{limma}}).
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
example_vsn <- norm_vsn(example_filter)
example_impute <- imputation(example_filter, fun = "MinProb", q = 0.01)

example_lm <- linear_model(example_impute, "Ctrl", "control")
}
