% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_se}
\alias{make_se}
\title{Data.frame to SummerizedExperiment using an experimental design}
\usage{
make_se(data, columns, expdesign)
}
\arguments{
\item{data}{Data.frame, Data object which will be turned into a SummerizedExperiment}

\item{columns}{Vector of integers, Number of the columns that contain the assay data.}

\item{expdesign}{Data.frame, Experimental design object}
}
\value{
A SummerizedExperiment object.
}
\description{
\code{make_se} creates a SummerizedExperiment object based on two data.frames: the data and experimental design.
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)
}
