% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_se_parse}
\alias{make_se_parse}
\title{Data.frame to SummarizedExperiment parsing from column names}
\usage{
make_se_parse(data, columns)
}
\arguments{
\item{data}{Data.frame, Data object which will be turned into a SummarizedExperiment}

\item{columns}{Vector of integers, Column numbers that contain the assay data.}
}
\value{
A SummarizedExperiment object.
}
\description{
\code{make_se_parse} creates a SummarizedExperiment object based on a single data.frame
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
example_se <- make_se_parse(example_unique, columns)
}
