% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_se_parse}
\alias{make_se_parse}
\title{Data.frame to SummarizedExperiment object
conversion using parsing from column names}
\usage{
make_se_parse(proteins_unique, columns)
}
\arguments{
\item{proteins_unique}{Data.frame,
Protein table with unique names annotated in the 'name' column.}

\item{columns}{Integer vector,
Column numbers indicating the columns containing the assay data.}
}
\value{
A SummarizedExperiment object
with log2-transformed values.
}
\description{
\code{make_se_parse} creates a SummarizedExperiment object
based on a single data.frame.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, columns)
}
