% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{manual_impute}
\alias{manual_impute}
\title{Imputation by random draws from a manually defined distribution}
\usage{
manual_impute(data, scale = 0.3, shift = 1.8)
}
\arguments{
\item{data}{SummarizedExperiment,
Proteomics data for which missing values will be imputed.}

\item{scale}{Numeric,
Sets the width of the distribution relative to the
standard deviation of the original distribution.}

\item{shift}{Numeric,
Sets the left-shift of the distribution in standard deviation from
the mean of the original distribution.}
}
\value{
An imputed SummarizedExperiment object.
}
\description{
\code{manual_impute} imputes missing values in a proteomics dataset
by random draws from a manually defined distribution.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)

imputed_manual <- impute(norm, fun = "man", shift = 1.8, scale = 0.3)
}
