% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ_functions.R
\name{merge_ibaq}
\alias{merge_ibaq}
\title{Merge iBAQ intensities}
\usage{
merge_ibaq(protein_unique, peptides)
}
\arguments{
\item{protein_unique}{Data.frame,
Protein table with unique names annotated in the 'name' column.}

\item{peptides}{Data.frame,
Peptides table from MaxQuant ('peptides.txt').}
}
\value{
A data.frame with iBAQ intensities per protein group.
Every protein group contains peptides unique to this group.
}
\description{
\code{merge_ibaq} generates a data.frame with merged iBAQ intensities
for proteins with shared peptides.
}
\examples{
data <- GFPip
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs')
peptides <- GFPip_pep

ibaq <- merge_ibaq(data_unique, peptides)
colnames(ibaq)
head(ibaq)
}
