% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{norm_vsn}
\alias{norm_vsn}
\title{Data normalization using vsn}
\usage{
norm_vsn(data)
}
\arguments{
\item{data}{SummerizedExperiment, Data object which will be normalized.}
}
\value{
An normalized SummerizedExperiment object.
}
\description{
\code{norm_vsn} nomralizes an SummerizedExperiment object using \code{\link[vsn]{vsn-package}}.
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
example_vsn <- norm_vsn(example_filter)
}
