% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalize_vsn}
\alias{normalize_vsn}
\title{Normalization using vsn}
\usage{
normalize_vsn(data)
}
\arguments{
\item{data}{SummarizedExperiment, Proteomics data with log2-transformed values (as data obtained from \code{\link{make_se}}.}
}
\value{
A normalized SummarizedExperiment object.
}
\description{
\code{normalize_vsn} performs normalization using \code{\link[vsn]{vsn-package}}.
}
\examples{
data <- UbiLength
data <- filter(data, Reverse != "+", Potential.contaminant != "+")
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
}
