% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions2.R
\name{plot_cond_freq}
\alias{plot_cond_freq}
\title{Plot frequency of significant conditions per protein}
\usage{
plot_cond_freq(data)
}
\arguments{
\item{data}{SummarizedExperiment,
Data object for which differentially enriched proteins are annotated
by \code{\link{test_diff}} and \code{\link{add_rejections}}.}
}
\value{
A histogram (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_cond_freq} generates a histogram of the number of significant conditions per protein.
}
\examples{
# load data
data <- UbiLength
data <- data[data$Reverse != '+' & data$Potential.contaminant != '+',]
exp_design <- UbiLength_ExpDesign

# perform the differential enrichment test
results <- LFQ(data, exp_design, "MinProb", "Ctrl", "control")
signif <- results$signif # get the significance object

# plot frequencies
plot_cond_freq(signif)

# or perform the step-by-step analysis
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs', delim = ';')
columns <- grep('LFQ.', colnames(data_unique))
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = 'MinProb', q = 0.01)

diff <- test_diff(imputed, 'Ctrl', 'control')
signif <- add_rejections(diff, alpha = 0.05, lfc = 1)

plot_cond_freq(signif)
}
