% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions2.R
\name{plot_corr}
\alias{plot_corr}
\title{Plot correlation matrix}
\usage{
plot_corr(data, significant = TRUE, lower = -1, upper = 1, pal = "PRGn",
  pal_rev = FALSE)
}
\arguments{
\item{data}{SummarizedExperiment,
Data object for which differentially enriched proteins are annotated
by \code{\link{test_diff}} and \code{\link{add_rejections}}.}

\item{significant}{Boolean,
Whether or not to filter for significant proteins.}

\item{lower}{Integer,
The lower limit of the color scale.}

\item{upper}{Integer,
The upper limit of the color scale.}

\item{pal}{Character,
The name of the color panel from \code{\link[RColorBrewer]{brewer.pal}}.}

\item{pal_rev}{Boolean,
Whether or not to invert the color palette.}
}
\value{
A heatmap plot (generated by \code{\link[ComplexHeatmap]{Heatmap}})
}
\description{
\code{plot_corr} generates a Pearson correlation matrix.
}
\examples{
# load data
data <- UbiLength
data <- data[data$Reverse != '+' & data$Potential.contaminant != '+',]
exp_design <- UbiLength_ExpDesign

# perform the differential enrichment test
results <- LFQ(data, exp_design, "MinProb", "Ctrl", "control")
signif <- results$signif # get the significance object

# plot correlation matrix
plot_corr(signif)
plot_corr(signif, lower = 0, upper = 1, pal = "Reds")

# or perform the step-by-step analysis
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs', delim = ';')
columns <- grep('LFQ.', colnames(data_unique))
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = 'MinProb', q = 0.01)

diff <- test_diff(imputed, 'Ctrl', 'control')
signif <- add_rejections(diff, alpha = 0.05, lfc = 1)

plot_corr(signif)
}
