% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot protein coverage in conditions}
\usage{
plot_coverage(data)
}
\arguments{
\item{data}{ExpressionSet, Data object for which to plot observation frequency.}
}
\value{
Barplot of protein coverage in samples (generated by \code{\link{ggplot}})
}
\description{
\code{plot_coverage} generates a barplot of the protein coverage in samples.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_se <- make_se(example_unique, columns, exp_design)
plot_coverage(example_se)

example_filter <- filter_missval(example_se, thr = 0)
plot_coverage(example_filter)
}
