% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_frequencies.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot protein coverage}
\usage{
plot_coverage(data)
}
\arguments{
\item{data}{SummarizedExperiment,
Data object for which to plot observation frequency.}
}
\value{
Barplot of protein coverage in samples
(generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_coverage} generates a barplot
of the protein coverage in all samples.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != '+' & data$Potential.contaminant != '+',]
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs', delim = ';')

columns <- grep('LFQ.', colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
plot_coverage(filt)
}
