% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot protein overlap between samples}
\usage{
plot_frequency(data)
}
\arguments{
\item{data}{ExpressionSet, Data object for which to plot observation frequency.}
}
\value{
Barplot of overlap of protein identifications between samples (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_frequency} generates a barplot of the protein overlap between samples
}
\examples{
data <- UbiLength
data <- filter(data, Reverse != "+", Potential.contaminant != "+")
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
plot_frequency(filt)
}
