% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot protein overlap between samples}
\usage{
plot_frequency(data)
}
\arguments{
\item{data}{ExpressionSet, Data object for which to plot observation frequency.}
}
\value{
Barplot of overlap of protein identifications between samples (generated by \code{\link{ggplot}})
}
\description{
\code{plot_frequency} generates a barplot of the protein overlap between samples
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)
plot_frequency(example_se)

example_filter <- filter_missval(example_se, thr = 0)
plot_frequency(example_filter)
}
