% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(data, type, k = 6, col_limit = 6, labelsize = 10)
}
\arguments{
\item{data}{SummarizedExperiment, Data object which has been generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}

\item{type}{"centered" or "contrast", The type of data scaling used for plotting.}

\item{k}{Integer, sets the number of k-means clusters.}

\item{col_limit}{Integer, sets the outer limits of the color scale.}

\item{labelsize}{Integer, sets the size of name labels.}
}
\value{
A heatmap (generated by \code{\link[ComplexHeatmap]{Heatmap}})
}
\description{
\code{plot_heatmap} generates a heatmap of all significant proteins.
}
\examples{
data <- UbiLength
data <- filter(data, Reverse != "+", Potential.contaminant != "+")
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.01)

diff <- test_diff(imputed, "Ctrl", "control")
signif <- add_rejections(diff, alpha = 0.05, lfc = 1)

plot_heatmap(signif, "centered", k = 6, col_limit = 4, labelsize = 3)
plot_heatmap(signif, "contrast", k = 6, col_limit = 10, labelsize = 3)
}
