% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(data, type, k = 6, col_limit = 6, labelsize = 10)
}
\arguments{
\item{data}{SummarizedExperiment, Data object which has been generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}

\item{type}{"centered" or "contrast", The type of data scaling used for plotting.}

\item{k}{Integer, sets the number of k-means clusters.}

\item{col_limit}{Integer, sets the outer limits of the color scale.}

\item{labelsize}{Integer, sets the size of name labels.}
}
\value{
A heatmap (generated by \code{\link[ComplexHeatmap]{Heatmap}})
}
\description{
\code{plot_heatmap} generates a heatmap of all significant proteins.
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
example_vsn <- norm_vsn(example_filter)
example_impute <- imputation(example_filter, fun = "MinProb", q = 0.01)

example_lm <- linear_model(example_impute, "Ctrl", "control")
example_sign <- cutoffs(example_lm, alpha = 0.05, lfc = 1)

plot_heatmap(example_sign, "centered", k = 6, col_limit = 4, labelsize = 3)
plot_heatmap(example_sign, "contrast", k = 6, col_limit = 10, labelsize = 3)
}
