% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_results.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(dep, type = c("contrast", "centered"), kmeans = FALSE, k = 6,
  col_limit = 6, row_font_size = 6, col_font_size = 10, ...)
}
\arguments{
\item{dep}{SummarizedExperiment,
Data object which has been generated by
\code{\link{test_diff}} and \code{\link{add_rejections}}.}

\item{type}{'contrast' or 'centered',
The type of data scaling used for plotting.}

\item{kmeans}{Logical(1),
Whether or not to perform k-means clustering.}

\item{k}{Integer(1),
Sets the number of k-means clusters.}

\item{col_limit}{Integer(1),
Sets the outer limits of the color scale.}

\item{row_font_size}{Integer(1),
Sets the size of row labels.}

\item{col_font_size}{Integer(1),
Sets the size of column labels.}

\item{...}{Additional arguments for Heatmap function as depicted in
\code{\link[ComplexHeatmap]{Heatmap}}}
}
\value{
A heatmap (generated by \code{\link[ComplexHeatmap]{Heatmap}})
}
\description{
\code{plot_heatmap} generates a heatmap of all significant proteins.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != '+' & data$Potential.contaminant != '+',]
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs', delim = ';')

columns <- grep('LFQ.', colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = 'MinProb', q = 0.01)

diff <- test_diff(imputed, 'Ctrl', 'control')
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)

plot_heatmap(dep)
plot_heatmap(dep, 'centered', kmeans = TRUE, k = 6, row_font_size = 3)
plot_heatmap(dep, 'contrast', col_limit = 10, row_font_size = 3)
}
