% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(data, type, k = 6, col_limit = 6)
}
\arguments{
\item{data}{ExpressionSet, Data object which has been generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}

\item{type}{"centered" or "contrast", The type of data scaling used for plotting.}

\item{k}{Integer, sets the number of k-means clusters.}

\item{col_limit}{Integer, sets the outer limits of the color scale.}
}
\value{
A heatmap generated by \code{\link{Heatmap}}
}
\description{
\code{plot_heatmap} generates a heatmap of all significant proteins.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
example_exprset <- into_exprset(example_unique, columns)
example_vsn <- norm_vsn(example_exprset)
example_filter <- miss_val_filter(example_vsn, thr = 0)
example_impute <- imputation_MSn(example_filter, fun = "QRILC")

example_lm <- linear_model(example_impute, "Con_", "control")
example_sign <- cutoffs(example_lm, alpha = 0.05, lfc = 1)

plot_heatmap(example_sign, "centered", k = 6, col_limit = 4)
plot_heatmap(example_sign, "contrast", k = 6, col_limit = 10)
}
