% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBAQ.R
\name{plot_ibaq}
\alias{plot_ibaq}
\title{Plot the iBAQ values versus log fold change}
\usage{
plot_ibaq(data, contrast, labelsize = 3)
}
\arguments{
\item{data}{SummarizedExperiment object, Proteomics dataset on which differential enrichment proteins are annotated by \code{\link{add_rejections}}.}

\item{contrast}{Character, The specific contrast to plot.}

\item{labelsize}{Integer, Sets the size of name labels.}
}
\value{
A scatter plot
}
\description{
\code{plot_ibaq} plots a scatter plot of the iBAQ intensities versus the LFQ fold changes.
}
\examples{
# load data
data <- GFPip
exp_design <- GFPip_ExpDesign
peptides <- GFPip_pep

# Make unique names and filter for reverse and contaminant hits
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs')
data_filt <- data_unique[data_unique$Reverse != '+' & data_unique$Contaminant != '+',]

# Make a SummarizedExperiment and perform filtering on missing values,
# variance stabilization and imputation on this object
cols <- grep('LFQ', colnames(data_filt))
se <- make_se(data_filt, cols, exp_design)
filt <- filter_missval(se)
norm <- normalize_vsn(filt)
imputed <- impute(norm, 'MinProb')

# Test for differential enriched proteins
diff <- test_diff(imputed, 'WT', 'control')
final <- add_rejections(diff, alpha = 0.05, lfc = 4.5)

# Plot iBAQ vs LFQ plot
plot_ibaq(final, 'GFP_vs_WT', labelsize = 3)

}
