% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_imputation}
\alias{plot_imputation}
\title{Visualize imputation}
\usage{
plot_imputation(raw, imp)
}
\arguments{
\item{raw}{SummarizedExperiment, Data object before imputation}

\item{imp}{SummarizedExperiment, Data object after imputation by \code{\link{impute}}.}
}
\value{
Density plots for all conditions before and after imputated (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_imputation} generates density plots for all conditions before and after imputation
}
\examples{
data <- UbiLength
data <- data[data$Reverse != '+' & data$Potential.contaminant != '+',]
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs', delim = ';')

columns <- grep('LFQ.', colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = 'MinProb', q = 0.01)

plot_imputation(norm, imputed)
}
