% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_missval}
\alias{plot_missval}
\title{Plot a heatmap of proteins with missing values}
\usage{
plot_missval(data)
}
\arguments{
\item{data}{SummerizedExperiment, Data object which has been generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}
}
\value{
A heatmap indicating whether values are missing (0) or not (1) (generated by \code{\link{Heatmap}})
}
\description{
\code{plot_missval} generates a heatmap of proteins with missing values to discover whether values are missing by random or not.
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
plot_missval(example_filter)
}
