% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_norm}
\alias{plot_norm}
\title{Visualize normalization}
\usage{
plot_norm(raw, norm)
}
\arguments{
\item{raw}{ExpressionSet, Data object before normalization.}

\item{norm}{ExpressionSet, Data object after normalization by \code{\link{norm_vsn}}.}
}
\value{
Boxplots for all samples before and after normalization generated by \code{\link{ggplot}}
}
\description{
\code{plot_norm} generates boxplots for all samples before and after normalization.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
example_exprset <- into_exprset(example_unique, columns)
example_vsn <- norm_vsn(example_exprset)

plot_norm(example_exprset, example_vsn)
}
