% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_QC.R
\name{plot_normalization}
\alias{plot_normalization}
\title{Visualize normalization}
\usage{
plot_normalization(raw, norm)
}
\arguments{
\item{raw}{SummarizedExperiment,
Data object before normalization.}

\item{norm}{SummarizedExperiment,
Data object after normalization by \code{\link{normalize_vsn}}.}
}
\value{
Boxplots for all conditions
before and after normalization (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_normalization} generates boxplots
for all conditions before and after normalization.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != '+' & data$Potential.contaminant != '+',]
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs', delim = ';')

columns <- grep('LFQ.', colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)

plot_normalization(filt, norm)
}
