% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_numbers}
\alias{plot_numbers}
\title{Plot protein numbers}
\usage{
plot_numbers(data)
}
\arguments{
\item{data}{SummerizedExperiment, Data object for which to plot protein numbers.}
}
\value{
Barplot of the number of identified proteins per condition (generated by \code{\link{ggplot}})
}
\description{
\code{plot_numbers} generates a barplot of the number of identified proteins per condition.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_exprset <- make_se(example_unique, columns, exp_design)
plot_numbers(example_exprset)

example_filter <- filter_missval(example_exprset, thr = 0)
plot_numbers(example_filter)
}
