% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_numbers}
\alias{plot_numbers}
\title{Plot protein numbers}
\usage{
plot_numbers(data)
}
\arguments{
\item{data}{ExpressionSet, Data object for which to plot protein numbers.}
}
\value{
Barplot of the number of identified proteins per samples generated by \code{\link{ggplots}}
}
\description{
\code{plot_numbers} generates a barplot of the number of identified proteins per samples.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
example_exprset <- into_exprset(example_unique, columns)
plot_numbers(example_exprset)

example_filter <- miss_val_filter(example_vsn, thr = 0)
plot_numbers(example_filter)
}
