% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_results.R
\name{plot_single}
\alias{plot_single}
\title{Plot values for a protein of interest}
\usage{
plot_single(dep, protein, type = c("contrast", "centered"))
}
\arguments{
\item{dep}{SummarizedExperiment,
Data object for which differentially enriched proteins are annotated
(output from \code{\link{test_diff}()} and \code{\link{add_rejections}()}).}

\item{protein}{Character(1),
Protein name.}

\item{type}{'contrast' or 'centered',
The type of data scaling used for plotting.}
}
\value{
A barplot (generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{plot_single} generates a barplot of a protein of interest.
}
\examples{
# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter, normalize and impute missing values
filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.01)

# Test for differentially expressed proteins
diff <- test_diff(imputed, "Ctrl", "control")
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)

# Plot single proteins
plot_single(dep, 'USP15', 'centered')
plot_single(dep, 'USP15', 'contrast')
}
