% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_results.R
\name{plot_single}
\alias{plot_single}
\title{Plot values for a protein of interest}
\usage{
plot_single(data, protein, type)
}
\arguments{
\item{data}{SummarizedExperiment,
Data object which has been generated
by \code{\link{test_diff}} and \code{\link{add_rejections}}.}

\item{protein}{Character,
Protein name.}

\item{type}{'centered' or 'contrast',
The type of data scaling used for plotting.}
}
\value{
A barplot (generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{plot_single} generates a barplot of a protein of interest.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != '+' & data$Potential.contaminant != '+',]
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs', delim = ';')

columns <- grep('LFQ.', colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = 'MinProb', q = 0.01)

diff <- test_diff(imputed, 'Ctrl', 'control')
signif <- add_rejections(diff, alpha = 0.05, lfc = 1)

plot_single(signif, 'USP15', 'centered')
plot_single(signif, 'USP15', 'contrast')
}
