% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions_results.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot}
\usage{
plot_volcano(dep, contrast, label_size = 3, add_names = TRUE)
}
\arguments{
\item{dep}{SummarizedExperiment,
Data object which has been generated by
\code{\link{test_diff}} and \code{\link{add_rejections}}.}

\item{contrast}{Character(1),
Specifies the contrast to plot.}

\item{label_size}{Integer(1),
Sets the size of name labels.}

\item{add_names}{Logical(1),
Whether or not to plot names.}
}
\value{
A volcano plot (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_volcano} generates a volcano plot for a specified contrast.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != '+' & data$Potential.contaminant != '+',]
data_unique <- make_unique(data, 'Gene.names', 'Protein.IDs', delim = ';')

columns <- grep('LFQ.', colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = 'MinProb', q = 0.01)

diff <- test_diff(imputed, 'Ctrl', 'control')
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)

plot_volcano(dep, 'Ubi6_vs_Ctrl', labelsize = 5, add_names = TRUE)
plot_volcano(dep, 'Ubi6_vs_Ctrl', add_names = FALSE)
plot_volcano(dep, 'Ubi4_vs_Ctrl', labelsize = 5, add_names = TRUE)
plot_volcano(dep, 'Ubi4_vs_Ctrl', add_names = FALSE)
}
