% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{results}
\alias{results}
\title{Generate a results table}
\usage{
results(data)
}
\arguments{
\item{data}{SummerizedExperiment, Data object which has been generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}
}
\value{
An data.frame object containing all results variables from the performed analysis.
}
\description{
\code{results} generates a results table (data.frame) from a SummerizedExperiment object which has been generated by \code{\link{linear_model}} followed by \code{\link{cutoffs}}.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- ExpDesign_UbIA_MS
example_se <- make_se(example_unique, columns, exp_design)

example_filter <- filter_missval(example_se, thr = 0)
example_vsn <- norm_vsn(example_filter)
example_impute <- imputation(example_filter, fun = "MinProb", q = 0.01)

example_lm <- linear_model(example_impute, "Ctrl", "control")
example_sign <- cutoffs(example_lm, alpha = 0.05, lfc = 1)
example_results <- results(example_sign)
glimpse(example_results)
}
