% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app_function.R
\name{run_app}
\alias{run_app}
\title{DEP shiny apps}
\usage{
run_app(app)
}
\arguments{
\item{app}{"LFQ", "TMT" or "iBAQ", The name of the app.}
}
\description{
\code{run_app} Launches an interactive shiny app for interactive analysis of proteomics data.
}
\examples{
\dontrun{
# Run the app
run_app("LFQ")

run_app("TMT")

run_app("iBAQ")

}
}
