% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{se2msn}
\alias{se2msn}
\title{Obtain a MSnSet object from a SummarizedExperiment object}
\usage{
se2msn(data)
}
\arguments{
\item{data}{SummarizedExperiment, Data object which will be turned into a MSnSet object.}
}
\value{
A MSnSet object.
}
\description{
\code{se2msn} generats a MSnSet object from a SummarizedExperiment object
}
\examples{
example <- UbiLength \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
exp_design <- UbiLength_ExpDesign
example_se <- make_se(example_unique, columns, exp_design)

example_MSnSet <- se2msn(example_se)
}
