% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{se2msn}
\alias{se2msn}
\title{SummarizedExperiment to MSnSet object conversion}
\usage{
se2msn(data)
}
\arguments{
\item{data}{SummarizedExperiment,
Data object which will be turned into a MSnSet object.}
}
\value{
A MSnSet object.
}
\description{
\code{se2msn} generats a MSnSet object from a SummarizedExperiment object.
}
\examples{
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

data_msn <- se2msn(se)
}
