% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_diff}
\alias{test_diff}
\title{Differential enrichment test}
\usage{
test_diff(data, control, type, test = NULL)
}
\arguments{
\item{data}{SummarizedExperiment,
Proteomics data which will be tested for differentially enriched proteins.}

\item{control}{Character,
The condition to which contrasts are generated
(a control condition would be most appropriate).}

\item{type}{"all", "control" or "manual",
The type of contrasts that will be tested.
This can be all possible pairwise comparisons ("all"),
limited to the comparisons versus the control ("control"), or
manually defined contrast(s) ("manual").}

\item{test}{Character,
The contrasts that will be tested if type = "manual".
These should be formatted as "SampleA_vs_SampleB" or
c("SampleA_vs_SampleC", "SampleB_vs_SampleC").}
}
\value{
An SummarizedExperiment object
containing FDR estimates of differential expression.
}
\description{
\code{test_diff} performs a differential enrichment test based on
protein-wise linear models and empherical Bayes
statistics (\code{\link[limma]{limma}}).
}
\examples{
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.01)

diff <- test_diff(imputed, "Ctrl", "control")
diff <- test_diff(imputed, "Ctrl", "manual",
    test = c("Ubi4_vs_Ctrl", "Ubi6_vs_Ctrl"))
}
