% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_diff}
\alias{test_diff}
\title{Differential enrichment test}
\usage{
test_diff(data, control, type)
}
\arguments{
\item{data}{SummarizedExperiment, Proteomics data which will be tested for differentially enriched proteins.}

\item{control}{Character, The condition to which contrasts are generated (a control condition would be most appropriate).}

\item{type}{"all" or "control", The type of contrasts that will be tested. Either all possible pairwise comparisons ("all") or limited to the comparisons versus the control ("control").}
}
\value{
An SummarizedExperiment object containing FDR estimates of differential expression.
}
\description{
\code{test_diff} performs a differential enrichment test based on protein-wise linear models and empherical Bayes statistics (\code{\link[limma]{limma}}).
}
\examples{
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.01)

diff <- test_diff(imputed, "Ctrl", "control")
}
