% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_theme.R
\name{theme_DEP1}
\alias{theme_DEP1}
\title{DEP ggplot theme 1}
\usage{
theme_DEP1()
}
\description{
\code{theme_DEP1} is the ggplot theme's used for plotting in \code{\link{DEP}} with horizontal x-axis labels. \code{theme_DEP2} has vertical x-axis labels.
}
