% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_functions.R
\name{TMT}
\alias{TMT}
\title{TMT workflow}
\usage{
TMT(proteins, expdesign, fun, control, type, test = NULL,
  name = "gene_name", ids = "protein_id", alpha = 0.05, lfc = 1)
}
\arguments{
\item{proteins}{Data.frame,
The data object.}

\item{expdesign}{Data.frame,
The experimental design object.}

\item{fun}{Character,
Function used for data imputation based on \code{\link[MSnbase]{impute}}.}

\item{control}{Character,
The sample name to which the contrasts are generated
(the control sample would be most appropriate).}

\item{type}{'all' or 'control',
The type of contrasts that will be generated.}

\item{test}{Character,
The contrasts that will be tested if type = "manual".
These should be formatted as "SampleA_vs_SampleB" or
c("SampleA_vs_SampleC", "SampleB_vs_SampleC").}

\item{name}{Character,
Name of the column representing gene names.}

\item{ids}{'Character,
Name of the column representing protein IDs.}

\item{alpha}{Numeric,
sets the false discovery rate threshold.}

\item{lfc}{Numeric,
sets the log fold change threshold.}
}
\value{
A list of 8 objects:
\item{se}{SummarizedExperiment object containing the original data}
\item{filt}{SummarizedExperiment object containing the filtered data}
\item{norm}{SummarizedExperiment object containing the normalized data}
\item{imputed}{SummarizedExperiment object containing the imputed data}
\item{diff}{SummarizedExperiment object
containing FDR estimates of differential expression}
\item{dep}{SummarizedExperiment object
annotated with logical columns indicating significant proteins}
\item{results}{data.frame containing containing
all results variables from the performed analysis}
\item{param}{data.frame containing the test parameters}
}
\description{
\code{TMT} is a wrapper function running the
entire differential enrichment/expression analysis workflow
for TMT-based proteomics data.
The protein table from
\href{http://www.nature.com/nprot/journal/v10/n10/full/nprot.2015.101.html}{IsobarQuant}
is used as direct input.
}
\examples{
\dontrun{

TMT_res <- TMT()

}
}
