% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UbiLength_ExpDesign}
\alias{UbiLength_ExpDesign}
\title{Experimental design of the UbiLength dataset}
\format{A data.frame with 12 observations and 3 variables:
\describe{
  \item{label}{Label names}
  \item{condition}{Experimental conditions}
  \item{replicate}{Replicate number}
}}
\source{
Zhang, Smits, van Tilburg, et al (2017). An interaction landscape of ubiquitin signaling. Molecular Cell:
  \url{http://www.cell.com/molecular-cell/fulltext/S1097-2765(17)30004-7}
}
\usage{
UbiLength_ExpDesign
}
\value{
A data.frame with the UbiLength Experimental design
}
\description{
This dataset annotates 12 different samples in 4 conditions and 3 replicates.
}
\keyword{datasets}
