% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{volcano}
\alias{volcano}
\title{Volcano plot}
\usage{
volcano(data, contrast, labelsize = 3, add_names = TRUE)
}
\arguments{
\item{data}{ExpressionSet, Data object which has been generated by \code{\link{linear_model}} and \code{\link{cutoffs}}.}

\item{contrast}{Character, The specific contrast to plot.}

\item{labelsize}{Integer, sets the size of name labels.}

\item{add_names}{Boolean, whether or not to plot names.}
}
\value{
A volcano plot generated by \code{\link{ggplot}}
}
\description{
\code{volcano} generates a volcano plot for a specified contrast.
}
\examples{
example <- UbIA_MS \%>\% filter(Reverse != "+", Potential.contaminant != "+")
example_unique <- unique_names(example, "Gene.names", "Protein.IDs", delim = ";")

columns <- grep("LFQ.", colnames(example_unique))
example_exprset <- into_exprset(example_unique, columns)
example_vsn <- norm_vsn(example_exprset)
example_filter <- miss_val_filter(example_vsn, thr = 0)
example_impute <- imputation_MSn(example_filter, fun = "QRILC")

example_lm <- linear_model(example_impute, "Con_", "control")
example_sign <- cutoffs(example_lm, alpha = 0.05, lfc = 1)

volcano(example_sign, "X6Ubi_.Con_", labelsize = 5, add_names = T)
volcano(example_sign, "X6Ubi_.Con_", add_names = F)
volcano(example_sign, "X4Ubi_.Con_", labelsize = 5, add_names = T)
volcano(example_sign, "X4Ubi_.Con_", add_names = F)
}
